+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 
+## 0.5.1 (2021-02-01)
+
+### Changed (1 change)
+- Print warning messages to /dev/stderr
+
+### Fixed (2 changes)
+- Fix `[[: not found` error in `install-binary.sh` script
+- Fix showing repos serving depreacted charts if `--ignore-deprecation` is set to `true` ([#36 (comment 768978347)](https://github.com/fabmation-gmbh/helm-whatup/issues/36\#issuecomment-768978347))
+
+
 ## 0.5.0 (2021-01-27)
 
 ### Added (5 changes)